[refactor] split test into a separate file
